shopping_cart_api/requirements.txt
django
django-debug-toolbar
mysqlclient
djangorestframework
drf-nested-routers
django-filter
